Add files via upload
interpolator_V1 is the initial MATLAB code that I created to interpolate the data from a flexible mesh onto a fixed mesh. It does not need any save  .m file to operate. But it needs the csv files related to flexible mesh to be in its folder. This was not a good code, but line 4 to 32 may be useful.